packages/gef/pyproject.toml
[project]
name = "evo-data-converters-gef"
description = "Python data converters for GEF to Evo geoscience objects"
version = "0.0.1"
requires-python = ">=3.10"
license-files = ["LICENSE.md"]
dynamic = ["readme"]
authors = [
    { name = "Seequent", email = "support@seequent.com" }
]

dependencies = [
    "evo-data-converters-common",
    "pandas",
    "pint-pandas",
    "pygef[plot]",
    "pyarrow>=17.0.0",
]

[project.urls]
Source = "https://github.com/SeequentEvo/evo-python-sdk"
Tracker = "https://github.com/SeequentEvo/evo-python-sdk/issues"
Homepage = "https://www.seequent.com/"
Documentation = "https://developer.seequent.com/"

[dependency-groups]
dev = [
    "pytest",
    "pytest-mock>=3.15.1",
]

[tool.ruff]
src = ["src", "tests"]
line-length = 120
[build-system]
requires = ["hatchling", "hatch-fancy-pypi-readme"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/evo"]

[tool.hatch.metadata.hooks.fancy-pypi-readme]
content-type = "text/markdown"

[[tool.hatch.metadata.hooks.fancy-pypi-readme.fragments]]
path = "README.md"

[[tool.hatch.metadata.hooks.fancy-pypi-readme.substitutions]]
# Literal TOML strings (single quotes) need no escaping of backslashes.
# Converts relative images to absolute links in PyPI
pattern = '!\[(.+?)\]\(((?!https?://)\S+?\.(?:jpe?g|png|gif))\)'
replacement = '![\1](https://github.com/SeequentEvo/evo-data-converters/raw/main/packages/gef/\g<2>)'

[[tool.hatch.metadata.hooks.fancy-pypi-readme.substitutions]]
# Literal TOML strings (single quotes) need no escaping of backslashes.
# Converts relative links to absolute links in PyPI
pattern = '\[(.+?)\]\(((?!https?://)\S+?)\)'
replacement = '[\1](https://github.com/SeequentEvo/evo-data-converters/tree/main/packages/gef/\g<2>)'
